# server/api/openapi.yaml
openapi: 3.0.3
info:
  title: Teraslack Messaging API
  version: 0.1.0
  description: Canonical HTTP API contract for the Teraslack server.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /agents:
    get:
      tags:
        - Agents
      operationId: listAgents
      summary: List manageable agents
      description: Returns agents owned directly by the current user or by workspaces where the current user is an owner or admin.
      responses:
        '200':
          description: Manageable agents.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentsCollection'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    post:
      tags:
        - Agents
      operationId: createAgent
      summary: Create an agent
      description: Creates a first-class agent user owned either by the current user or by one workspace the current user administers.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAgentRequest'
      responses:
        '201':
          description: Agent created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
  /agents/{agent_id}:
    get:
      tags:
        - Agents
      operationId: getAgent
      summary: Show an agent
      description: Returns one manageable agent by its user identifier.
      parameters:
        - $ref: '#/components/parameters/AgentIDPath'
      responses:
        '200':
          description: Agent details.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
    patch:
      tags:
        - Agents
      operationId: updateAgent
      summary: Update an agent
      description: Updates a manageable agent's profile, mode, or status.
      parameters:
        - $ref: '#/components/parameters/AgentIDPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAgentRequest'
      responses:
        '200':
          description: Updated agent.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        default:
          $ref: '#/components/responses/ApiErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  responses:
    ApiErrorResponse:
      description: Error response.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  parameters:
...
    AgentIDPath:
      in: path
      name: agent_id
      required: true
      description: Agent user identifier.
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Stable machine-readable error code.
        message:
          type: string
          description: Human-readable error message.
        request_id:
          type: string
          description: Request identifier for server-side troubleshooting.
        errors:
          type: array
          description: Field-level validation details when request validation fails.
          items:
            $ref: '#/components/schemas/ValidationDetail'
    ValidationDetail:
      type: object
      required:
        - field
        - code
        - message
      properties:
        field:
          type: string
          description: Request field that failed validation.
        code:
          type: string
          description: Stable machine-readable validation error code.
        message:
          type: string
          description: Human-readable validation message.
...
    UserProfile:
      type: object
      required:
        - handle
        - display_name
      properties:
        handle:
          type: string
          description: Public user handle.
        display_name:
          type: string
          description: Human-friendly display name.
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: Avatar image URL.
        bio:
          type: string
          nullable: true
          description: Optional profile bio.
    User:
      type: object
      required:
        - id
        - principal_type
        - status
        - profile
      properties:
        id:
          type: string
          format: uuid
          description: User identifier.
        principal_type:
          type: string
          description: Principal type for the user record.
        status:
          type: string
          description: User status.
        email:
          type: string
          format: email
          nullable: true
          description: Primary email address when available.
        profile:
          $ref: '#/components/schemas/UserProfile'
    Agent:
      type: object
      required:
        - user
        - owner_type
        - mode
        - created_by_user_id
        - created_at
        - updated_at
      properties:
        user:
          $ref: '#/components/schemas/User'
        owner_type:
          type: string
          enum:
            - user
            - workspace
          description: Whether the agent is directly owned by a user or by one workspace.
        owner_user_id:
          type: string
          format: uuid
          nullable: true
          description: Direct owner when owner_type is user.
        owner_workspace_id:
          type: string
          format: uuid
          nullable: true
          description: Owning workspace when owner_type is workspace.
        mode:
          type: string
          enum:
            - read_only
            - safe_write
          description: Effective write mode for API-key-authenticated agent actions.
        created_by_user_id:
          type: string
          format: uuid
          description: User who created the agent.
        created_at:
          type: string
          format: date-time
          description: Creation time.
        updated_at:
          type: string
          format: date-time
          description: Last metadata update time.
    AgentsCollection:
      type: object
      required:
        - items
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        next_cursor:
          type: string
          description: Pagination cursor when additional results are available.
    CreateAgentRequest:
      type: object
      required:
        - display_name
        - owner_type
      properties:
        display_name:
          type: string
          description: Agent display name.
        handle:
          type: string
          nullable: true
          description: Optional public handle. When omitted, the server derives one from display_name.
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: Optional avatar image URL.
        bio:
          type: string
          nullable: true
          description: Optional profile bio.
        owner_type:
          type: string
          enum:
            - user
            - workspace
          description: Ownership model for the new agent.
        owner_workspace_id:
          type: string
          format: uuid
          nullable: true
          description: Required when owner_type is workspace. Must be omitted for user-owned agents.
        mode:
          type: string
          enum:
            - read_only
            - safe_write
          default: safe_write
          description: Initial write mode for the agent. Defaults to safe_write when omitted.
    UpdateAgentRequest:
      type: object
      properties:
        display_name:
          type: string
          description: Updated display name.
        handle:
          type: string
          description: Updated public handle.
        avatar_url:
          type: string
          format: uri
          nullable: true
          description: Updated avatar URL or null to clear it.
        bio:
          type: string
          nullable: true
          description: Updated bio or null to clear it.
        mode:
          type: string
          enum:
            - read_only
            - safe_write
          description: Updated write mode.
        status:
          type: string
          enum:
            - active
            - suspended
          description: Updated user status for the agent.
...